pyproject.toml
[project]
name = "dbt-forge"
dynamic = ["version"]
description = "Scaffold production-ready dbt projects with opinionated defaults"
readme = { file = "README.md", content-type = "text/markdown" }
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.11,<3.14"
authors = [
    { name = "Marouane" },
]
maintainers = [
    { name = "Marouane" },
]
keywords = [
    "analytics-engineering",
    "cli",
    "code-generator",
    "dbt",
    "scaffolding",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Utilities",
]
urls = { Homepage = "https://github.com/maroil/dbt-forge", Repository = "https://github.com/maroil/dbt-forge", Issues = "https://github.com/maroil/dbt-forge/issues", Changelog = "https://github.com/maroil/dbt-forge/blob/main/CHANGELOG.md" }
dependencies = [
    "jinja2>=3.1.6",
    "questionary>=2.1.1",
    "rich>=14.3.3",
    "typer>=0.12",
]

[project.scripts]
dbt-forge = "dbt_forge.main:app"

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/dbt_forge/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
    "/CONTRIBUTING.md",
    "/CODE_OF_CONDUCT.md",
    "/SECURITY.md",
    "/pyproject.toml",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dbt_forge"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
major_version_zero = true

[dependency-groups]
dev = [
    "build>=1.2.2",
    "commitizen>=4.9.1",
    "pre-commit>=4.3.0",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "ruff>=0.13.0",
    "twine>=6.1.0",
]
